+++ src/components/AddTodoForm.jsx
@@ -1,5 +1,7 @@
 import { useState } from 'react';
 import InputWithLabel from './InputWithLabel';
+import styles from './AddTodoForm.module.css';
+import PropTypes from 'prop-types';
 
 const AddTodoForm = ({ onAddTodo }) => {
   const [todoTitle, setTodoTitle] = useState('');
@@ -11,7 +13,7 @@ const AddTodoForm = ({ onAddTodo }) => {
   const handleAddTodo = (event) => {
     event.preventDefault();
     if (!todoTitle.trim()) {
-      alert('Please enter a valid todo title.');
+      alert("Please enter a valid todo title.");
       return;
     }
     if (onAddTodo) {
@@ -20,19 +22,25 @@ const AddTodoForm = ({ onAddTodo }) => {
         id: Date.now().toString(),
       });
     }
-    setTodoTitle(''); // Reset input
+    setTodoTitle(""); // Reset input
   };
 
   return (
-    <form onSubmit={handleAddTodo}>
-      <InputWithLabel
-        label="Title"
-        value={todoTitle}
-        onChange={handleTitleChange}
-      />
+    <form onSubmit={handleAddTodo} className={styles.form}>
+      <div className={styles.inputGroup}>
+        <InputWithLabel
+          label="Title"
+          value={todoTitle}
+          onChange={handleTitleChange}
+        />
+      </div>
       <button type="submit">Add</button>
     </form>
   );
+};
+
+AddTodoForm.propTypes = {
+  onAddTodo: PropTypes.func.isRequired,
 };
 
 export default AddTodoForm;
